feat: optimize the package name for cross-domain functionality (#690)
