+++ backend/crossdomain/contract/agent/single_agent.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package crossagent
+package agent
 
 import (
 	"context"
