+++ backend/crossdomain/contract/message/message.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package crossmessage
+package message
 
 import (
 	"context"
