+++ backend/domain/agent/singleagent/internal/agentflow/node_tool_database.go
@@ -33,7 +33,7 @@ import (
 	"github.com/coze-dev/coze-studio/backend/api/model/app/bot_common"
 	"github.com/coze-dev/coze-studio/backend/api/model/crossdomain/database"
 	"github.com/coze-dev/coze-studio/backend/api/model/data/database/table"
-	"github.com/coze-dev/coze-studio/backend/crossdomain/contract/crossdatabase"
+	crossdatabase "github.com/coze-dev/coze-studio/backend/crossdomain/contract/database"
 	"github.com/coze-dev/coze-studio/backend/domain/memory/database/service"
 	"github.com/coze-dev/coze-studio/backend/infra/impl/sqlparser"
 	"github.com/coze-dev/coze-studio/backend/pkg/lang/slices"
