+++ backend/domain/agent/singleagent/internal/agentflow/node_tool_knowledge.go
@@ -28,7 +28,7 @@ import (
 	"github.com/getkin/kin-openapi/openapi3"
 
 	"github.com/coze-dev/coze-studio/backend/api/model/app/bot_common"
-	"github.com/coze-dev/coze-studio/backend/crossdomain/contract/crossknowledge"
+	crossknowledge "github.com/coze-dev/coze-studio/backend/crossdomain/contract/knowledge"
 	knowledgeEntity "github.com/coze-dev/coze-studio/backend/domain/knowledge/entity"
 )
