+++ backend/domain/app/service/service_impl.go
@@ -25,10 +25,10 @@ import (
 	connectorModel "github.com/coze-dev/coze-studio/backend/api/model/crossdomain/connector"
 	databaseModel "github.com/coze-dev/coze-studio/backend/api/model/crossdomain/database"
 	knowledgeModel "github.com/coze-dev/coze-studio/backend/api/model/crossdomain/knowledge"
-	"github.com/coze-dev/coze-studio/backend/crossdomain/contract/crossconnector"
-	"github.com/coze-dev/coze-studio/backend/crossdomain/contract/crossdatabase"
-	"github.com/coze-dev/coze-studio/backend/crossdomain/contract/crossknowledge"
-	"github.com/coze-dev/coze-studio/backend/crossdomain/contract/crossplugin"
+	crossconnector "github.com/coze-dev/coze-studio/backend/crossdomain/contract/connector"
+	crossdatabase "github.com/coze-dev/coze-studio/backend/crossdomain/contract/database"
+	crossknowledge "github.com/coze-dev/coze-studio/backend/crossdomain/contract/knowledge"
+	crossplugin "github.com/coze-dev/coze-studio/backend/crossdomain/contract/plugin"
 	"github.com/coze-dev/coze-studio/backend/domain/app/entity"
 	"github.com/coze-dev/coze-studio/backend/domain/app/repository"
 	"github.com/coze-dev/coze-studio/backend/infra/contract/idgen"
